Enhance error messages in payment state management and refactor package structure for clarity and Kafka communication

+++ src/main/java/com/sap/adds_service/prices/application/usecases/deleteprice/DeletePriceCase.java
@@ -2,6 +2,8 @@ package com.sap.adds_service.prices.application.usecases.deleteprice;
 
 import com.sap.adds_service.prices.application.input.DeletePriceCasePort;
 import com.sap.adds_service.prices.application.output.DeletePricePort;
+import com.sap.adds_service.prices.application.output.FindPricePort;
+import com.sap.common_lib.exception.NotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,14 +16,21 @@ import java.util.UUID;
 public class DeletePriceCase implements DeletePriceCasePort {
 
     private final DeletePricePort deletePricePort;
+    private final FindPricePort findPricePort;
 
     @Override
     public void deleteByCinemaId(UUID cinemaId) {
+        if (!findPricePort.checkIfCinemaExistsById(cinemaId)) {
+            throw new NotFoundException("Price not found for cinema id: " + cinemaId);
+        }
         deletePricePort.deleteByCinemaId(cinemaId);
     }
 
     @Override
     public void deleteById(UUID id) {
+        findPricePort.findById(id).orElseThrow(
+                () -> new NotFoundException("Price not found for id: " + id)
+        );
         deletePricePort.deleteById(id);
     }
 }
